Introduces a lightweight, more efficient way of validating translations.
Instead of building HTML, this update changes the target to plaintext. This significantly reduces the overhead associated with HTML rendering, resulting in a more efficient validation process. On my device, it reduces around 85% of the validation time.

Additionally, for local translations, this commit introduces an incremental build, replacing the full build. This change reduces the validation process time from approximately 20 minutes to just a few seconds on a local device.